requirements.txt
black==24.10.0
Django==5.1.2
django-stubs==5.1.1
django-stubs-ext==5.1.1 
djangorestframework==3.15.2
djangorestframework-stubs==3.15.1
flake8==7.1.1
mypy==1.13.0
mypy-extensions==1.0.0
mysqlclient==2.2.5
pillow==11.0.0
random-username==1.0.2
django-resized==1.0.3
django-redis==5.4.0
PyJWT==2.9.0
social_auth_app_django==5.4.2
pyyaml==6.0.2 
ua-parser==0.18.0 
user-agents==2.2.0
django-hide-admin==1.0.0
django-appconf==1.0.6
django-select2==8.2.1
django-mdeditor==0.1.0
Markdown==3.7